bodge add video output
you need to add ffmpeg.exe to the same folder as onnxUI.py.

this will let you set a range of steps to then convert into a video. this can take a while as its generating a full image for each step value you set. with high resolutions it could take hours for a large range so be warned.

this should work with upwards or downwards steps, and also should work for txt2img, inpainting, & img2img.